Change soundVolume from float to int, add extra args to popup data, make centering multiline text possible, fix retarded compilation errors with C<->C++

+++ globals.h
@@ -1,3 +1,3 @@
 #pragma once
 
-extern float soundVolume;
+extern int soundVolume;
